fix(idea-controllers) No se crean dos ideas con el mismo nombre - FormularioIdea.php

+++ controllers/FormularioIdea.php
@@ -116,44 +116,51 @@ EOF;
         if(empty($result)){
           try{
         	$idea = new Idea;
-        	$idea->setNombre_Idea($datos['nombre']);
-        	$idea->setId_Categoria($datos['categoria']);
-        	$idea->setFecha_Limite($datos['fecha']);
-        	$idea->setDesc_idea($desc);
-        	$idea->setEnVenta($datos['enventa']);
-          $idea->setImporte_Solicitado($datos['recaudacion']);
-        	$idea->setImporte_venta($datos['precio']);
-        	$idea->setId_Correo($_SESSION['mail']);
+			$consulta = $idea->getBy("nombre_idea",$datos['nombre']); 
+			$nombreIdea= $consulta ? $consulta[0]['nombre_idea']:null;
+			$existe=strcasecmp($nombreIdea,$datos['nombre']);
+			if($existe!=0){
+				$idea->setNombre_Idea($datos['nombre']);
+				$idea->setId_Categoria($datos['categoria']);
+				$idea->setFecha_Limite($datos['fecha']);
+				$idea->setDesc_idea($desc);
+				$idea->setEnVenta($datos['enventa']);
+			  $idea->setImporte_Solicitado($datos['recaudacion']);
+				$idea->setImporte_venta($datos['precio']);
+				$idea->setId_Correo($_SESSION['mail']);
 
-        	$idea->setIdea();
+				$idea->setIdea();
 
-          $newdir = "../images/ideas/idea".$idea->getId_idea()."/";
-          //Comprueba que la imagen sea un archivo de imagen
-          $image_file = $newdir . basename($_FILES["foto"]["name"]);
-          $imageFileType = strtolower(pathinfo($image_file,PATHINFO_EXTENSION));
+			  $newdir = "../images/ideas/idea".$idea->getId_idea()."/";
+			  //Comprueba que la imagen sea un archivo de imagen
+			  $image_file = $newdir . basename($_FILES["foto"]["name"]);
+			  $imageFileType = strtolower(pathinfo($image_file,PATHINFO_EXTENSION));
 
-          if($imageFileType != "jpg" && $imageFileType != "png" && $imageFileType != "jpeg" && $imageFileType != "gif" )
-              $result[] = "La imagen debe tener extensión: ( .jpg | .png | .jpeg | .gif )";
+			  if($imageFileType != "jpg" && $imageFileType != "png" && $imageFileType != "jpeg" && $imageFileType != "gif" )
+				  $result[] = "La imagen debe tener extensión: ( .jpg | .png | .jpeg | .gif )";
 
-          //Comprueba que el curriculum sea un pdf
-          $curr_file = $newdir . basename($_FILES["archivo"]["name"]);
-          $currFileType = strtolower(pathinfo($curr_file,PATHINFO_EXTENSION));
+			  //Comprueba que el curriculum sea un pdf
+			  $curr_file = $newdir . basename($_FILES["archivo"]["name"]);
+			  $currFileType = strtolower(pathinfo($curr_file,PATHINFO_EXTENSION));
 
-          if($currFileType != "pdf") $result[] = "El curriculum debe tener extensión .pdf";
+			  if($currFileType != "pdf") $result[] = "El curriculum debe tener extensión .pdf";
 
-          if (!file_exists($newdir) && !is_dir($newdir))
-            mkdir($newdir, 0777);
-          //Sube la Imagen
-          if (!move_uploaded_file($_FILES["foto"]["tmp_name"], $image_file))
-              $result[]="Error al subir la imagen";
+			  if (!file_exists($newdir) && !is_dir($newdir))
+				mkdir($newdir, 0777);
+			  //Sube la Imagen
+			  if (!move_uploaded_file($_FILES["foto"]["tmp_name"], $image_file))
+				  $result[]="Error al subir la imagen";
 
-          //Sube el cv
-          if (!move_uploaded_file($_FILES["archivo"]["tmp_name"], $curr_file))
-              $result[]="Error al subir el Curriculum";
+			  //Sube el cv
+			  if (!move_uploaded_file($_FILES["archivo"]["tmp_name"], $curr_file))
+				  $result[]="Error al subir el Curriculum";
 
-      		$idea->setFiles($curr_file,$image_file);
-      	  $idea->closeConnection();
-          $result = "../views/infoIdea.php?id_idea=".$idea->getId_idea();
+				$idea->setFiles($curr_file,$image_file);
+			  $idea->closeConnection();
+				$result = "../views/infoIdea.php?id_idea=".$idea->getId_idea();
+			}else{
+				$result[]="Lo sentimos, esa idea ya existe";
+			}
       	}catch(Exception $e){
       		error_log("MySQL: Code: ".$e->getCode(). " Desc: " .$e->getMessage() ,0);
       		$_SESSION['data_error']=$e->getMessage();
